+++ pages/shops/[shopId]/index.tsx
@@ -1,11 +1,12 @@
 import Button from "@/components/Button/Button";
 import Layout from "@/components/Layout";
-import PostForm from "@/components/PostForm/PostForm";
+import PostForm from "@/components/PostForm/index";
 import { useToast } from "@/contexts/ToastContext";
 import { h1 } from "@/styles/fontsStyle";
 import styled from "@emotion/styled";
 import { useRouter } from "next/router";
 import { useState } from "react";
+import handleRegisterClick from "./utils/handleRegisterClick";
 
 export default function NoticeRegistrationPage() {
   const router = useRouter();
@@ -25,26 +26,6 @@ export default function NoticeRegistrationPage() {
     }));
   };
 
-  const handleRegisterClick = () => {
-    if (
-      formData.wage.trim() === "" ||
-      formData.start_at.trim() === "" ||
-      formData.work_time.trim() === "" ||
-      formData.description.trim() === ""
-    ) {
-      showToast("모든 필드를 작성해주세요.");
-      return;
-    }
-
-    const confirmed = window.confirm(
-      `\n시급: ${formData.wage}\n시작 일시: ${formData.start_at}\n업무 시간: ${formData.work_time}\n공고 설명: ${formData.description}\n\n등록하시겠습니까?`,
-    );
-    if (confirmed) {
-      router.push("/"); // 공고 등록 한 후 id를 이용해 shops/${shopId}/notices/${noticeId}로 이동해야함
-      showToast("등록되었습니다!");
-    }
-  };
-
   return (
     <Layout>
       <Wrapper>
@@ -54,7 +35,9 @@ export default function NoticeRegistrationPage() {
           text="등록하기"
           color="colored"
           width={312}
-          handleClick={handleRegisterClick}
+          handleClick={() => {
+            handleRegisterClick(formData, showToast, router);
+          }}
         />
       </Wrapper>
     </Layout>
