+++ pages/shops/[shopId]/types/types.ts
@@ -1,0 +1,6 @@
+export interface FormDataType {
+  wage: string;
+  start_at: string;
+  work_time: string;
+  description: string;
+}
